Add files via upload
The main idea is the same as the idea of protection in FACEBOOK. In FACEBOOK, when
the user enters the incorrect password for six times for example, FACEBOOK can detect if
the user is a person or a robot, but how?
FACEBOOK suggests that it is a robot guessing the password as it calculates the time
between each password the user enter. If this time is less than 0.5 sec for example, it is a
robot else it’s a person.
This project also detects if the user is a person or a robot using the same principle.